+++ packages/server/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@repo/server",
   "version": "0.0.1",
-  "type": "mudule",
   "scripts": {
     "dev": "wrangler dev src/index.ts",
     "deploy": "wrangler deploy --minify src/index.ts",
