+++ packages/server/src/routes/auth/index.ts
@@ -1,10 +1,7 @@
-// import { app } from "../../base/base-app";
-import { OpenAPIHono } from "@hono/zod-openapi";
+import { app } from "../../base/base-app";
 import { loginHandler, loginSpec } from "./login";
 import { registerHandler, registerSpec } from "./register";
 
-export const authRouter = new OpenAPIHono({
-	strict: false,
-})
+export const authRouter = app
 	.openapi(loginSpec, loginHandler)
 	.openapi(registerSpec, registerHandler);
