+++ packages/server/src/routes/auth/logout.ts
@@ -1,109 +1,30 @@
 import { createRoute, RouteHandler, z } from "@hono/zod-openapi";
-import { loginFields } from "@repo/zod-types";
-import { ENV_TYPES } from "../../env/zod";
-import { env } from "hono/adapter";
-import { AuthTable } from "./handlers";
-import { checkPassword } from "../../utils/bcrypt";
-import { generateToken } from "../../jwt_token";
 import { setCookie } from "hono/cookie";
 import { COOKIES } from "../../env/cookies";
 
-export const loginSpec = createRoute({
-	method: "post",
-	path: "/auth/login",
+export const logoutSpec = createRoute({
+	method: "delete",
+	path: "/auth/logout",
 	tags: ["auth"],
 	responses: {
-		201: {
+		200: {
 			description: "Register Successful",
 			content: {
 				"application/json": {
 					schema: z.object({
-						data: z.object({
-							name: z.string(),
-							email: z.string(),
-							phone: z.string().nullable(),
-							id: z.string(),
-							emailConfirmedAt: z.string().nullable(),
-							deletedAt: z.string().nullable(),
-							createdAt: z.string(),
-							updateAt: z.string().nullable(),
-						}),
+						data: z.literal("Logged out"),
 					}),
-				},
-			},
-		},
-		400: {
-			description: "Bad Request",
-			content: {
-				"application/json": {
-					schema: z.object({
-						error: z.string(),
-					}),
-				},
-			},
-		},
-		403: {
-			description: "Bad Request",
-			content: {
-				"application/json": {
-					schema: z.object({
-						error: z.string(),
-					}),
-				},
-			},
-		},
-		404: {
-			description: "Not Found",
-			content: {
-				"application/json": {
-					schema: z.object({
-						error: z.string(),
-					}),
-				},
-			},
-		},
-	},
-	request: {
-		body: {
-			content: {
-				"application/json": {
-					schema: loginFields,
 				},
 			},
 		},
 	},
 });
 
-type LoginRoute = typeof loginSpec;
+type LogoutRoute = typeof logoutSpec;
 
-export const loginHandler: RouteHandler<LoginRoute> = async (c) => {
-	const form = c.req.valid("json");
-	const {
-		TURSO_AUTH_TOKEN,
-		TURSO_CONNECTION_URL,
-		COOkIE_SECRET_KEY,
-		JWT_SECRET_KEY,
-	} = env<ENV_TYPES>(c);
+export const logoutHandler: RouteHandler<LogoutRoute> = async (c) => {
+	setCookie(c, COOKIES.USER_ID, "", { maxAge: 0 });
+	setCookie(c, COOKIES.USER_TOKEN, "", { maxAge: 0 });
 
-	if (!form) return c.json({ error: "Invalid form data" }, 400);
-
-	const { email, password } = form;
-
-	const Auth = new AuthTable(TURSO_CONNECTION_URL, TURSO_AUTH_TOKEN);
-	const user = await Auth.findAuthUser(email);
-	if (!user) return c.json({ error: "User not found" }, 404);
-
-	const doesPasswordMatch = await checkPassword(password, user.passwordDigest);
-
-	if (!doesPasswordMatch)
-		return c.json({ error: "Password does not match" }, 403);
-
-	const token = await generateToken(user, JWT_SECRET_KEY);
-
-	setCookie(c, COOKIES.USER_ID, user.email);
-	setCookie(c, COOKIES.USER_TOKEN, token);
-
-	const { passwordDigest: NOT_USE, ...rest } = user;
-
-	return c.json({ data: rest }, 201);
+	return c.json({ data: "Logged out" }, 200);
 };
