+++ packages/zod-types/process/index.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
 
 export const envVariables = z.object({
-	TURSO_CONNECTION_URL: z.string().default(""),
-	TURSO_AUTH_TOKEN: z.string().default(""),
-	JWT_SECRET_KEY: z.string().default(""),
-	COOkIE_SECRET_KEY: z.string().default(""),
+	TURSO_CONNECTION_URL: z.string(),
+	TURSO_AUTH_TOKEN: z.string(),
+	JWT_SECRET_KEY: z.string(),
+	COOkIE_SECRET_KEY: z.string(),
 	LANG: z.string().default("pt_BR"),
 });
